Fix crasher in S3Aggregator
This crash results when the records for a site are flushed by
S3Aggregator::save_batch_if_past_timeout (thus deleting the raw records
for that visit_id). If no new records are saved for that visit_id and a
new page is visted by the crawler, the aggregator attempts to flush the
same record batch before moving to the next site. We can skip that
second attempt at a flush.