Remove --user flag from pip install
Remove the --user flag from 'pip install' commands. This flag caused
install.sh and install-dev.sh to fail when run in a virtual environment,
because user site-packages were not visible. It is no longer necessary,
as the current default behavior of pip is to perform a user install when
not running in a virtual environment and the global site-packages
directory is not writeable.